Issue 394 Issue 252
394: Replace code change lost with move of DAmail MailService.java to
K-9.  Eliminate the stopSelfs.  Without the stopSelfs, the Service
remains active and the Android platform is far less likely to kill the
process, although not absolutely forbidden from doing so.  Better will
be to implement the full BroadcastReceiver methodology referred to in
Issue 394, which would allow replacing the stopSelfs, which would make
K-9 a better good citizen.

252: Fix logic for raising and canceling notifications based on number
of new unread messages and total unread messages.



git-svn-id: http://k9mail.googlecode.com/svn/k9mail/trunk@468 9a7fa126-a48b-11dd-a887-7ffe1a420f8d